refactor(#14): move common types to shared files
