+++ common/api/GetPaginationInfo.ts
@@ -1,0 +1,10 @@
+export type GetPaginationInfoRequest = {
+	itemsOnPage: number;
+	tag?: string;
+	lang?: string;
+};
+
+export type GetPaginationInfoResponse = {
+	postsNumber: number;
+	pagesNumber: number;
+};
