+++ common/api/GetPost.ts
@@ -1,0 +1,1 @@
+export type GetPostRequest = { url: string };
