+++ common/api/GetPostWithAdditionalData.ts
@@ -1,0 +1,1 @@
+export type GetPostWithAdditionalDataRequest = { url: string };
