+++ common/api/GetPosts.ts
@@ -1,0 +1,7 @@
+export type GetPostsRequest = {
+	from?: number;
+	limit?: number;
+	tag?: string;
+	lang?: string;
+	sort?: 'asc' | 'desc';
+};
