+++ server/api/getPost.ts
@@ -1,15 +1,11 @@
 import { ApiKnob } from '../types';
-import { buildApiRequest } from '../lib/api';
 
-import { getPost as getPostFn, Post } from '../lib/posts';
-
-export type Params = { url: string };
-
-export const getPost = buildApiRequest<Params, Post>('/getPost');
+import { getPost as getPostFn } from '../lib/posts';
+import { GetPostRequest } from '../../common/api/GetPost';
 
 export const getPostFabric: ApiKnob = (app) => {
 	app.get('/getPost', async function(req, res) {
-		const { url } = req.query as Params;
+		const { url } = req.query as GetPostRequest;
 
 		const post = await getPostFn(url);
