+++ server/api/getPosts.ts
@@ -1,21 +1,11 @@
 import { ApiKnob } from '../types';
-import { buildApiRequest } from '../lib/api';
 
-import { getPosts as getPostsFn, Post } from '../lib/posts';
-
-export type Params = {
-	from?: number;
-	limit?: number;
-	tag?: string;
-	lang?: string;
-	sort?: 'asc' | 'desc';
-};
-
-export const getPosts = buildApiRequest<Params, Post[]>('/getPosts');
+import { getPosts as getPostsFn } from '../lib/posts';
+import { GetPostsRequest } from '../../common/api/GetPosts';
 
 export const getPostsFabric: ApiKnob = (app) => {
 	app.get('/getPosts', async function(req, res) {
-		const options = req.query as Params;
+		const options = req.query as GetPostsRequest;
 
 		const posts = await getPostsFn(options);
