+++ server/lib/posts/post.ts
@@ -1,4 +1,3 @@
-import { MDXRemoteSerializeResult } from 'next-mdx-remote';
 import { serialize } from 'next-mdx-remote/serialize';
 import { visit } from 'unist-util-visit';
 import getReadingTime from 'reading-time';
@@ -9,45 +8,11 @@ import remarkRehype from 'remark-rehype';
 import rehypeStringify from 'rehype-stringify';
 import rehypeHighlight from 'rehype-highlight';
 
-import { readFile, stat } from 'fs/promises';
-
-import { extractTimestampFromName, getPostUrlByFilename } from './files';
 import path from 'path';
+import { readFile } from 'fs/promises';
 
-export type Post = {
-	url: string;
-	date: number;
-
-	title: string;
-	description: string | null;
-	previewText: string;
-	tags: string[];
-	keywords: string[];
-	lang: string | null;
-
-	/**
-	 * Image extracted from post to insert in meta tags
-	 * Prefer to `previewImage` if specified
-	 */
-	coverImage: string | null;
-
-	/**
-	 * Image specified manually in metadata
-	 */
-	previewImage: string | null;
-
-	source: MDXRemoteSerializeResult<Record<string, unknown>>;
-	readingTime: {
-		minutes: number;
-		words: number;
-	};
-};
-
-export type PostWithAdditionalData = Post & {
-	additionalData: {
-		html: string;
-	};
-};
+import { PostWithAdditionalData } from '../../../common/Post';
+import { extractTimestampFromName, getPostUrlByFilename } from './files';
 
 export const isUrl = (filename: string) => /^[a-z]*:\/\//i.test(filename);
 export const isRelativePath = (filename: string) => /^[^\/]/.test(filename);
