+++ site/api/api.ts
@@ -1,13 +1,15 @@
-import { stringify } from 'query-string';
+import queryString from 'query-string';
 
-import { OptionalKeys } from '../types';
-import { apiPath } from '../constants';
+import { OptionalKeys } from '../../server/types';
+import { apiPath } from '../../server/constants';
 
 export const buildApiRequest =
 	<T extends Record<string, any> | void = void, R = void>(path: string) =>
 		(options: keyof T extends OptionalKeys<T> ? T | void : T) => {
 			const url =
-			apiPath + path + (options === undefined ? '' : '?' + stringify(options));
+			apiPath +
+			path +
+			(options === undefined ? '' : '?' + queryString.stringify(options));
 
 			return fetch(url).then((r) => r.json() as Promise<R>);
 		};
