+++ site/api/requests/getPaginationInfo.ts
@@ -1,0 +1,10 @@
+import { buildApiRequest } from '../api';
+import {
+	GetPaginationInfoRequest,
+	GetPaginationInfoResponse,
+} from '../../../common/api/GetPaginationInfo';
+
+export const getPaginationInfo = buildApiRequest<
+	GetPaginationInfoRequest,
+	GetPaginationInfoResponse
+>('/getPaginationInfo');
