+++ site/api/requests/getPostUrls.ts
@@ -1,0 +1,3 @@
+import { buildApiRequest } from '../api';
+
+export const getPostUrls = buildApiRequest<void, string[]>('/getPostUrls');
