+++ site/api/requests/getPostWithAdditionalData.ts
@@ -1,0 +1,8 @@
+import { buildApiRequest } from '../api';
+import { PostWithAdditionalData } from '../../../common/Post';
+import { GetPostWithAdditionalDataRequest } from '../../../common/api/GetPostWithAdditionalData';
+
+export const getPostWithAdditionalData = buildApiRequest<
+	GetPostWithAdditionalDataRequest,
+	PostWithAdditionalData
+>('/getPostWithAdditionalData');
