+++ site/api/requests/getPosts.ts
@@ -1,0 +1,5 @@
+import { buildApiRequest } from '../api';
+import { Post } from '../../../common/Post';
+import { GetPostsRequest } from '../../../common/api/GetPosts';
+
+export const getPosts = buildApiRequest<GetPostsRequest, Post[]>('/getPosts');
